refactor: styled component 중복파일 삭제
